refactor(studio): extract builder shell concerns into hooks
Builder was a 743-line orchestrator mixing unrelated concerns. Pull the
self-contained shell effects into focused, reusable hooks:

- useForceDarkTheme — forces dark on <html> while the studio is mounted
- useStudioInit — one-time seed (stashed fork project / ?component= deep-link)
- useStudioLayout — persisted panel layout (load/save + hydration-safe key)

Behavior is unchanged; builder.tsx drops to 668 lines and the logic is now
independently testable.
