+++ apps/web/lib/generated-sources.ts
@@ -341,8 +341,8 @@ export const compositionSources: Record<string, CompositionSource> = {
   },
   TweetCard: {
     component:
-      '"use client";\nimport {\n  AiImageIcon,\n  ArrowDown01Icon,\n  Calendar03Icon,\n  CheckListIcon,\n  EarthIcon,\n  Flag02Icon,\n  Gif02Icon,\n  Image02Icon,\n  Location01Icon,\n  SmileIcon,\n} from "@hugeicons/core-free-icons";\nimport { HugeiconsIcon } from "@hugeicons/react";\nimport { AbsoluteFill, Img } from "remotion";\nimport { type ClipStyle, resolveClipStyle } from "../../clip-style";\nimport { proxyExternalImg } from "../../proxy-image";\nimport { useDesignFrame } from "../../use-design-frame";\n\nexport type TweetCardProps = {\n  displayName: string;\n  avatarUrl: string;\n  text: string;\n  audience: string;\n  theme: "light" | "dark";\n  clipStyle?: ClipStyle;\n};\n\nconst CARD_ENTER_END = 18;\nconst TYPE_START = 22;\nconst FRAMES_PER_CHAR = 2;\nconst PRESS_GAP = 16;\nconst PRESS_DURATION = 12;\n\nexport const TweetCard: React.FC<TweetCardProps> = ({\n  displayName,\n  avatarUrl,\n  text,\n  audience,\n  theme,\n  clipStyle,\n}) => {\n  const frame = useDesignFrame();\n\n  const s = resolveClipStyle(clipStyle, {\n    background: theme === "dark" ? "#000000" : "#ffffff",\n    color: theme === "dark" ? "#e7e9ea" : "#0f1419",\n    fontFamily:\n      "-apple-system, BlinkMacSystemFont, \'SF Pro Display\', Inter, sans-serif",\n    accent: "#1d9bf0",\n  });\n\n  const typingDuration = text.length * FRAMES_PER_CHAR;\n  const typingEnd = TYPE_START + typingDuration;\n  const charsTyped =\n    frame < TYPE_START\n      ? 0\n      : Math.min(text.length, Math.floor((frame - TYPE_START) / FRAMES_PER_CHAR));\n  const visibleText = text.slice(0, charsTyped);\n  const hasText = charsTyped > 0;\n  const isTyping = frame >= TYPE_START && frame < typingEnd;\n  // Caret blinks before typing starts and while idle; solid while typing.\n  const caretBlink = isTyping ? true : Math.floor(frame / 16) % 2 === 0;\n  const showCaret = frame >= CARD_ENTER_END && frame < typingEnd + 8;\n\n  // "Hit Post" click — a brief scale dip after typing completes.\n  const pressStart = typingEnd + PRESS_GAP;\n  const pressT =\n    frame < pressStart ? 0 : Math.min(1, (frame - pressStart) / PRESS_DURATION);\n  const postScale = 1 - Math.sin(pressT * Math.PI) * 0.12;\n  const posted = frame >= pressStart + PRESS_DURATION;\n\n  const isDark = theme === "dark";\n  const cardBg = isDark ? "#000000" : "#ffffff";\n  const cardText = isDark ? "#e7e9ea" : s.color;\n  const placeholder = isDark ? "#71767b" : "#536471";\n  const divider = isDark ? "#2f3336" : "#eff3f4";\n  const pillBorder = isDark ? "#36393b" : "#cfd9de";\n  const accent = s.accent;\n\n  // Post button: dimmed pill until there is text, then solid + bold.\n  const postBg = hasText\n    ? isDark\n      ? "#eff3f4"\n      : accent\n    : isDark\n      ? "#22282b"\n      : "#9bd1f9";\n  const postColor = hasText\n    ? isDark\n      ? "#0f1419"\n      : "#ffffff"\n    : isDark\n      ? "#6b7174"\n      : "#ffffff";\n\n  const toolbarIcons = [\n    Image02Icon,\n    Gif02Icon,\n    AiImageIcon,\n    CheckListIcon,\n    SmileIcon,\n    Calendar03Icon,\n    Location01Icon,\n    Flag02Icon,\n  ];\n\n  return (\n    <AbsoluteFill\n      style={{\n        background: s.background,\n        fontFamily: s.fontFamily,\n        display: "flex",\n        alignItems: "center",\n        justifyContent: "center",\n        padding: 80,\n      }}\n    >\n      <div\n        style={{\n          width: 1480,\n          background: cardBg,\n          color: cardText,\n          borderRadius: 32,\n          border: `1px solid ${divider}`,\n          boxShadow: isDark\n            ? "0 40px 100px rgba(0,0,0,0.5)"\n            : "0 36px 100px rgba(15,16,20,0.10), 0 6px 16px rgba(15,16,20,0.05)",\n          padding: 56,\n        }}\n      >\n        <div style={{ display: "flex", alignItems: "flex-start", gap: 28 }}>\n          <Avatar url={avatarUrl} initial={displayName.slice(0, 1)} />\n\n          <div style={{ flex: 1, minWidth: 0, paddingTop: 4 }}>\n            {/* Audience pill */}\n            <div\n              style={{\n                display: "inline-flex",\n                alignItems: "center",\n                gap: 8,\n                color: accent,\n                fontSize: 28,\n                fontWeight: 700,\n                border: `1px solid ${pillBorder}`,\n                borderRadius: 999,\n                padding: "8px 18px",\n              }}\n            >\n              <span>{audience}</span>\n              <HugeiconsIcon icon={ArrowDown01Icon} size={26} color={accent} />\n            </div>\n\n            {/* Compose area */}\n            <div\n              style={{\n                fontSize: 60,\n                fontWeight: 400,\n                lineHeight: 1.3,\n                letterSpacing: "-0.01em",\n                margin: "30px 0 0",\n                minHeight: 80,\n                color: hasText ? cardText : placeholder,\n                whiteSpace: "pre-wrap",\n                wordBreak: "break-word",\n              }}\n            >\n              {hasText ? visibleText : "What’s happening?"}\n              {showCaret && (\n                <span\n                  style={{\n                    display: "inline-block",\n                    width: 4,\n                    height: 62,\n                    marginLeft: hasText ? 6 : 2,\n                    background: accent,\n                    opacity: caretBlink ? 1 : 0,\n                    verticalAlign: "text-bottom",\n                    borderRadius: 2,\n                  }}\n                />\n              )}\n            </div>\n\n            {/* Everyone can reply */}\n            <div\n              style={{\n                display: "flex",\n                alignItems: "center",\n                gap: 12,\n                marginTop: 30,\n                color: accent,\n                fontSize: 30,\n                fontWeight: 700,\n              }}\n            >\n              <HugeiconsIcon icon={EarthIcon} size={32} color={accent} />\n              <span>Everyone can reply</span>\n            </div>\n          </div>\n        </div>\n\n        <div style={{ height: 1, background: divider, margin: "40px 0 28px" }} />\n\n        <div\n          style={{\n            display: "flex",\n            justifyContent: "space-between",\n            alignItems: "center",\n          }}\n        >\n          <div style={{ display: "flex", alignItems: "center", gap: 34 }}>\n            {toolbarIcons.map((icon, i) => (\n              <HugeiconsIcon\n                // eslint-disable-next-line react/no-array-index-key\n                key={i}\n                icon={icon}\n                size={40}\n                color={accent}\n              />\n            ))}\n          </div>\n\n          <div\n            style={{\n              background: postBg,\n              color: postColor,\n              fontSize: 34,\n              fontWeight: 700,\n              borderRadius: 999,\n              padding: "18px 46px",\n              transform: `scale(${postScale})`,\n              transformOrigin: "center",\n              boxShadow:\n                posted && isDark\n                  ? "0 0 0 6px rgba(239,243,244,0.12)"\n                  : "none",\n            }}\n          >\n            Post\n          </div>\n        </div>\n      </div>\n    </AbsoluteFill>\n  );\n};\n\nfunction Avatar({ url, initial }: { url: string; initial: string }) {\n  if (url.trim()) {\n    return (\n      <Img\n        src={proxyExternalImg(url)}\n        crossOrigin="anonymous"\n        width={92}\n        height={92}\n        style={{\n          width: 92,\n          height: 92,\n          borderRadius: "50%",\n          flexShrink: 0,\n          objectFit: "cover",\n        }}\n      />\n    );\n  }\n  return (\n    <div\n      style={{\n        width: 92,\n        height: 92,\n        borderRadius: "50%",\n        background: "linear-gradient(135deg, #1d9bf0 0%, #4f46e5 100%)",\n        color: "#ffffff",\n        display: "flex",\n        alignItems: "center",\n        justifyContent: "center",\n        fontSize: 40,\n        fontWeight: 700,\n        flexShrink: 0,\n      }}\n    >\n      {initial.toUpperCase() || "?"}\n    </div>\n  );\n}\n',
-    meta: 'import type { CompositionInfo } from "../../schema";\nimport type { TweetCardProps } from "./TweetCard";\n\nexport const TWEET_CARD_DURATION = 140;\nexport const TWEET_CARD_FPS = 60;\nexport const TWEET_CARD_WIDTH = 1920;\nexport const TWEET_CARD_HEIGHT = 1080;\n\nexport const tweetCardDefaultProps: TweetCardProps = {\n  displayName: "sanku",\n  avatarUrl: "https://avatars.githubusercontent.com/sankalpaacharya?s=200",\n  text: "just shipped something new 🚀",\n  audience: "Everyone",\n  theme: "dark",\n};\n\nexport const tweetCardInfo: CompositionInfo<TweetCardProps> = {\n  id: "TweetCard",\n  category: "social",\n  brandMode: "locked",\n  agentNotes:\n    "Authentic X / Twitter compose box: avatar, audience pill, \'What\'s happening?\' field, reply-scope line, the media toolbar, and a Post button. The tweet text types itself out and the Post button lights up and gets \'clicked\'. Use for \'we just posted this\' / announcement beats. Keep text under ~240 chars.",\n  title: "Tweet Composer",\n  description:\n    "An animated X / Twitter compose box. The tweet text types out and the Post button activates and is pressed.",\n  durationInFrames: TWEET_CARD_DURATION,\n  fps: TWEET_CARD_FPS,\n  width: TWEET_CARD_WIDTH,\n  height: TWEET_CARD_HEIGHT,\n  defaultProps: tweetCardDefaultProps,\n  fields: [\n    { kind: "text", key: "displayName", label: "Display name" },\n    { kind: "text", key: "avatarUrl", label: "Avatar URL" },\n    { kind: "textarea", key: "text", label: "Tweet text", rows: 3 },\n    { kind: "text", key: "audience", label: "Audience" },\n    {\n      kind: "select",\n      key: "theme",\n      label: "Theme",\n      options: [\n        { value: "light", label: "Light" },\n        { value: "dark", label: "Dark" },\n      ],\n    },\n  ],\n};\n',
+      '"use client";\nimport {\n  AiImageIcon,\n  ArrowDown01Icon,\n  Calendar03Icon,\n  CheckListIcon,\n  EarthIcon,\n  Flag02Icon,\n  Gif02Icon,\n  Image02Icon,\n  Location01Icon,\n  SmileIcon,\n} from "@hugeicons/core-free-icons";\nimport { HugeiconsIcon } from "@hugeicons/react";\nimport { AbsoluteFill, Img } from "remotion";\nimport { type ClipStyle, resolveClipStyle } from "../../clip-style";\nimport { proxyExternalImg } from "../../proxy-image";\nimport { useDesignFrame } from "../../use-design-frame";\n\nexport type TweetCardProps = {\n  displayName: string;\n  avatarUrl: string;\n  text: string;\n  audience: string;\n  theme: "light" | "dark";\n  clipStyle?: ClipStyle;\n};\n\nconst CARD_ENTER_END = 18;\nconst TYPE_START = 22;\nconst FRAMES_PER_CHAR = 2;\nconst PRESS_GAP = 16;\nconst PRESS_DURATION = 12;\n\nexport const TweetCard: React.FC<TweetCardProps> = ({\n  displayName,\n  avatarUrl,\n  text,\n  audience,\n  theme,\n  clipStyle,\n}) => {\n  const frame = useDesignFrame();\n\n  const s = resolveClipStyle(clipStyle, {\n    background: theme === "dark" ? "#000000" : "#ffffff",\n    color: theme === "dark" ? "#e7e9ea" : "#0f1419",\n    fontFamily:\n      "-apple-system, BlinkMacSystemFont, \'SF Pro Display\', Inter, sans-serif",\n    accent: "#1d9bf0",\n  });\n\n  const typingDuration = text.length * FRAMES_PER_CHAR;\n  const typingEnd = TYPE_START + typingDuration;\n  const charsTyped =\n    frame < TYPE_START\n      ? 0\n      : Math.min(\n          text.length,\n          Math.floor((frame - TYPE_START) / FRAMES_PER_CHAR),\n        );\n  const visibleText = text.slice(0, charsTyped);\n  const hasText = charsTyped > 0;\n  const isTyping = frame >= TYPE_START && frame < typingEnd;\n  // Caret blinks before typing starts and while idle; solid while typing.\n  const caretBlink = isTyping ? true : Math.floor(frame / 16) % 2 === 0;\n  const showCaret = frame >= CARD_ENTER_END && frame < typingEnd + 8;\n\n  // "Hit Post" click — a brief scale dip after typing completes.\n  const pressStart = typingEnd + PRESS_GAP;\n  const pressT =\n    frame < pressStart ? 0 : Math.min(1, (frame - pressStart) / PRESS_DURATION);\n  const postScale = 1 - Math.sin(pressT * Math.PI) * 0.12;\n  const posted = frame >= pressStart + PRESS_DURATION;\n\n  const isDark = theme === "dark";\n  const cardBg = isDark ? "#000000" : "#ffffff";\n  const cardText = isDark ? "#e7e9ea" : s.color;\n  const placeholder = isDark ? "#71767b" : "#536471";\n  const divider = isDark ? "#2f3336" : "#eff3f4";\n  const pillBorder = isDark ? "#36393b" : "#cfd9de";\n  const accent = s.accent;\n\n  // Post button: dimmed pill until there is text, then solid + bold.\n  const postBg = hasText\n    ? isDark\n      ? "#eff3f4"\n      : accent\n    : isDark\n      ? "#22282b"\n      : "#9bd1f9";\n  const postColor = hasText\n    ? isDark\n      ? "#0f1419"\n      : "#ffffff"\n    : isDark\n      ? "#6b7174"\n      : "#ffffff";\n\n  const toolbarIcons = [\n    Image02Icon,\n    Gif02Icon,\n    AiImageIcon,\n    CheckListIcon,\n    SmileIcon,\n    Calendar03Icon,\n    Location01Icon,\n    Flag02Icon,\n  ];\n\n  return (\n    <AbsoluteFill\n      style={{\n        background: s.background,\n        fontFamily: s.fontFamily,\n        display: "flex",\n        alignItems: "center",\n        justifyContent: "center",\n        padding: 80,\n      }}\n    >\n      <div\n        style={{\n          width: 1480,\n          background: cardBg,\n          color: cardText,\n          borderRadius: 32,\n          border: `1px solid ${divider}`,\n          boxShadow: isDark\n            ? "0 40px 100px rgba(0,0,0,0.5)"\n            : "0 36px 100px rgba(15,16,20,0.10), 0 6px 16px rgba(15,16,20,0.05)",\n          padding: 56,\n        }}\n      >\n        <div style={{ display: "flex", alignItems: "flex-start", gap: 28 }}>\n          <Avatar url={avatarUrl} initial={displayName.slice(0, 1)} />\n\n          <div style={{ flex: 1, minWidth: 0, paddingTop: 4 }}>\n            {/* Audience pill */}\n            <div\n              style={{\n                display: "inline-flex",\n                alignItems: "center",\n                gap: 8,\n                color: accent,\n                fontSize: 28,\n                fontWeight: 700,\n                border: `1px solid ${pillBorder}`,\n                borderRadius: 999,\n                padding: "8px 18px",\n              }}\n            >\n              <span>{audience}</span>\n              <HugeiconsIcon icon={ArrowDown01Icon} size={26} color={accent} />\n            </div>\n\n            {/* Compose area */}\n            <div\n              style={{\n                fontSize: 60,\n                fontWeight: 400,\n                lineHeight: 1.3,\n                letterSpacing: "-0.01em",\n                margin: "30px 0 0",\n                minHeight: 80,\n                color: hasText ? cardText : placeholder,\n                whiteSpace: "pre-wrap",\n                wordBreak: "break-word",\n              }}\n            >\n              {hasText ? visibleText : "What’s happening?"}\n              {showCaret && (\n                <span\n                  style={{\n                    display: "inline-block",\n                    width: 4,\n                    height: 62,\n                    marginLeft: hasText ? 6 : 2,\n                    background: accent,\n                    opacity: caretBlink ? 1 : 0,\n                    verticalAlign: "text-bottom",\n                    borderRadius: 2,\n                  }}\n                />\n              )}\n            </div>\n\n            {/* Everyone can reply */}\n            <div\n              style={{\n                display: "flex",\n                alignItems: "center",\n                gap: 12,\n                marginTop: 30,\n                color: accent,\n                fontSize: 30,\n                fontWeight: 700,\n              }}\n            >\n              <HugeiconsIcon icon={EarthIcon} size={32} color={accent} />\n              <span>Everyone can reply</span>\n            </div>\n          </div>\n        </div>\n\n        <div\n          style={{ height: 1, background: divider, margin: "40px 0 28px" }}\n        />\n\n        <div\n          style={{\n            display: "flex",\n            justifyContent: "space-between",\n            alignItems: "center",\n          }}\n        >\n          <div style={{ display: "flex", alignItems: "center", gap: 34 }}>\n            {toolbarIcons.map((icon, i) => (\n              <HugeiconsIcon\n                // eslint-disable-next-line react/no-array-index-key\n                key={i}\n                icon={icon}\n                size={40}\n                color={accent}\n              />\n            ))}\n          </div>\n\n          <div\n            style={{\n              background: postBg,\n              color: postColor,\n              fontSize: 34,\n              fontWeight: 700,\n              borderRadius: 999,\n              padding: "18px 46px",\n              transform: `scale(${postScale})`,\n              transformOrigin: "center",\n              boxShadow:\n                posted && isDark ? "0 0 0 6px rgba(239,243,244,0.12)" : "none",\n            }}\n          >\n            Post\n          </div>\n        </div>\n      </div>\n    </AbsoluteFill>\n  );\n};\n\nfunction Avatar({ url, initial }: { url: string; initial: string }) {\n  if (url.trim()) {\n    return (\n      <Img\n        src={proxyExternalImg(url)}\n        crossOrigin="anonymous"\n        width={92}\n        height={92}\n        style={{\n          width: 92,\n          height: 92,\n          borderRadius: "50%",\n          flexShrink: 0,\n          objectFit: "cover",\n        }}\n      />\n    );\n  }\n  return (\n    <div\n      style={{\n        width: 92,\n        height: 92,\n        borderRadius: "50%",\n        background: "linear-gradient(135deg, #1d9bf0 0%, #4f46e5 100%)",\n        color: "#ffffff",\n        display: "flex",\n        alignItems: "center",\n        justifyContent: "center",\n        fontSize: 40,\n        fontWeight: 700,\n        flexShrink: 0,\n      }}\n    >\n      {initial.toUpperCase() || "?"}\n    </div>\n  );\n}\n',
+    meta: 'import type { CompositionInfo } from "../../schema";\nimport type { TweetCardProps } from "./TweetCard";\n\nexport const TWEET_CARD_DURATION = 140;\nexport const TWEET_CARD_FPS = 60;\nexport const TWEET_CARD_WIDTH = 1920;\nexport const TWEET_CARD_HEIGHT = 1080;\n\nexport const tweetCardDefaultProps: TweetCardProps = {\n  displayName: "sanku",\n  avatarUrl: "https://avatars.githubusercontent.com/sankalpaacharya?s=200",\n  text: "just shipped something new 🚀",\n  audience: "Everyone",\n  theme: "dark",\n};\n\nexport const tweetCardInfo: CompositionInfo<TweetCardProps> = {\n  id: "TweetCard",\n  category: "social",\n  agentNotes:\n    "Authentic X / Twitter compose box: avatar, audience pill, \'What\'s happening?\' field, reply-scope line, the media toolbar, and a Post button. The tweet text types itself out and the Post button lights up and gets \'clicked\'. Use for \'we just posted this\' / announcement beats. Keep text under ~240 chars.",\n  title: "Tweet Composer",\n  description:\n    "An animated X / Twitter compose box. The tweet text types out and the Post button activates and is pressed.",\n  durationInFrames: TWEET_CARD_DURATION,\n  fps: TWEET_CARD_FPS,\n  width: TWEET_CARD_WIDTH,\n  height: TWEET_CARD_HEIGHT,\n  defaultProps: tweetCardDefaultProps,\n  fields: [\n    { kind: "text", key: "displayName", label: "Display name" },\n    { kind: "text", key: "avatarUrl", label: "Avatar URL" },\n    { kind: "textarea", key: "text", label: "Tweet text", rows: 3 },\n    { kind: "text", key: "audience", label: "Audience" },\n    {\n      kind: "select",\n      key: "theme",\n      label: "Theme",\n      options: [\n        { value: "light", label: "Light" },\n        { value: "dark", label: "Dark" },\n      ],\n    },\n  ],\n};\n',
   },
   TwitterFollow: {
     component:
